# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  backend_internal_models.Property:
    properties:
      county:
        type: string
      created_at:
        type: string
      date_of_transfer:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      district:
        type: string
      duration:
        description: F = Freehold, L = Leasehold
        type: string
      id:
        description: A unique identifier for each property sale
        type: string
      locality:
        type: string
      old_new:
        description: Y = a newly built property, N = an established residential building
        type: string
      paon:
        description: Primary Addressable Object Name. Typically the house number or
          name.
        type: string
      postcode:
        type: string
      ppd_category_type:
        description: A = Standard Price Paid entry, B = Additional Price Paid entry
        type: string
      price:
        description: In GBP
        type: integer
      property_type:
        description: D = Detached, S = Semi-Detached, T = Terraced, F = Flats/Maisonettes,
          O = Other
        type: string
      record_status:
        description: A = Addition, C = Change, D = Delete
        type: string
      saon:
        description: Secondary Addressable Object Name. Where a property has been
          divided into separate units (for example, flats).
        type: string
      street:
        type: string
      town_city:
        type: string
      updated_at:
        type: string
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  internal_routes_api.BaseResponse:
    properties:
      message:
        type: string
      status:
        example: true
        type: boolean
    type: object
  internal_routes_api.ErrorResponse:
    properties:
      error:
        type: string
      status:
        example: false
        type: boolean
    type: object
...
  internal_routes_api.PropertyListPayload:
    properties:
      data:
        items:
          $ref: '#/definitions/backend_internal_models.Property'
        type: array
      total:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a comprehensive API for analyzing and accessing UK housing
    market and rental data.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: UK Housing Market & Rental Intelligence API
  version: "1.0"
paths:
...
  /property:
    get:
      consumes:
      - application/json
      description: Get a list of UK housing properties with pagination and filtering
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Town/City
        in: query
        name: town_city
        type: string
      - description: District
        in: query
        name: district
        type: string
      - description: County
        in: query
        name: county
        type: string
      - description: Property Type
        in: query
        name: property_type
        type: string
      - description: Minimum Price
        in: query
        name: min_price
        type: integer
      - description: Maximum Price
        in: query
        name: max_price
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_routes_api.PropertyListPayload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_routes_api.ErrorResponse'
      summary: List properties
      tags:
      - property
    post:
      consumes:
      - application/json
      description: Create a new property record
      parameters:
      - description: Property object
        in: body
        name: property
        required: true
        schema:
          $ref: '#/definitions/backend_internal_models.Property'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/backend_internal_models.Property'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_routes_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_routes_api.ErrorResponse'
      security:
      - JwtAuth: []
      summary: Create property
      tags:
      - property
  /property/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a property record
      parameters:
      - description: Property ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_routes_api.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_routes_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_routes_api.ErrorResponse'
      security:
      - JwtAuth: []
      summary: Delete property
      tags:
      - property
    get:
      consumes:
      - application/json
      description: Get detailed information for a specific property
      parameters:
      - description: Property ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_models.Property'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_routes_api.ErrorResponse'
      summary: Get property by ID
      tags:
      - property
    put:
      consumes:
      - application/json
      description: Update an existing property record
      parameters:
      - description: Property ID
        in: path
        name: id
        required: true
        type: string
      - description: Property object
        in: body
        name: property
        required: true
        schema:
          $ref: '#/definitions/backend_internal_models.Property'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_models.Property'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_routes_api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_routes_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_routes_api.ErrorResponse'
      security:
      - JwtAuth: []
      summary: Update property
      tags:
      - property
securityDefinitions:
  JwtAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"